[PAM-2352] Legger til tekst "Ingen treff på" +  søkekriterer man har huket av
