+++ src/search/noResults/NoResults.less
@@ -2,6 +2,6 @@
 
 .NoResults {
   background: #fff;
-  padding: 4rem;
+  padding: 4rem 1rem;
   text-align: center;
 }
